# docs/tech/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pereobuyka API
  description: |
    HTTP API ядра «Переобуйка». Сценарии — docs/vision.md §4.
    LLM не подменяет факты; контекст подмешивается на сервере (vision §6).
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /api/v1/admin/dashboard/today:
    get:
      tags: [Admin]
      summary: KPI дашборда на сегодня
      operationId: adminDashboardToday
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTodayResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/dashboard/week-grid:
    get:
      tags: [Admin]
      summary: Сетка недели для записей и визитов
      operationId: adminWeekGrid
      security:
        - BearerAuth: []
      parameters:
        - name: week_start
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Понедельник недели
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekGridResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/analytics/week:
    get:
      tags: [Admin]
      summary: Аналитика недели для графиков
      operationId: adminAnalyticsWeek
      security:
        - BearerAuth: []
      parameters:
        - name: week_start
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsWeekResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/analytics/data-insight:
    post:
      tags: [Admin]
      summary: Вопрос к операционным данным (NL→SQL, только SELECT)
      description: >
        Генерация безопасного SELECT через LLM, валидация AST и whitelist таблиц (ADR-006).
        Требуется OPENROUTER_API_KEY на сервере.
      operationId: adminAnalyticsDataInsight
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDataInsightRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDataInsightResponse"
        "400":
          description: Запрос отклонён валидатором или ошибка выполнения SQL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: LLM недоступен или неверный ответ модели
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/clients:
    get:
      tags: [Admin]
      summary: Список клиентов (KPI)
      operationId: adminListClients
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClientListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/clients/{user_id}:
    get:
      tags: [Admin]
      summary: Карточка клиента
      operationId: adminGetClient
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminClientRow"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/users/{user_id}/appointments:
    get:
      tags: [Admin]
      summary: Записи выбранного пользователя
      operationId: adminUserAppointments
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAppointmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/users/{user_id}/visits:
    get:
      tags: [Admin]
      summary: Визиты выбранного пользователя
      operationId: adminUserVisits
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/appointments/{appointment_id}:
    patch:
      tags: [Admin]
      summary: Правка записи администратором
      operationId: adminPatchAppointment
      security:
        - BearerAuth: []
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAppointmentPatchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DomainError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/visits/{visit_id}:
    patch:
      tags: [Admin]
      summary: Правка визита администратором
      operationId: adminPatchVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVisitPatchBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/visits/{visit_id}/client-rating:
    post:
      tags: [Admin]
      summary: Оценка клиента администратором
      operationId: adminPostClientRating
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitRatingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/admin/appointments:
    get:
      tags: [Admin]
      summary: Журнал записей и загрузка
      operationId: adminListAppointments
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAppointmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Admin]
      summary: Создать запись на клиента
      description: |
        Источник записи на стороне сервера — `admin`. Требование к `starts_at`: не в прошлом
        (см. `STARTS_AT_IN_PAST` в `POST /api/v1/appointments`).
      operationId: adminCreateAppointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAppointmentCreateBody"
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Валидация или STARTS_AT_IN_PAST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Нет или невалидный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Ошибка валидации тела или параметров
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DomainError:
      description: Доменная ошибка (конфликт, нарушение бизнес-правила)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    UserRole:
      type: string
      enum: [client, admin]
    UserSource:
      type: string
      enum: [telegram, web]
    User:
      type: object
      required: [id, name, role, registered_at, source]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        telegram_id:
          type: integer
          format: int64
          nullable: true
        telegram_username:
          type: string
          nullable: true
        registered_at:
          type: string
          format: date-time
        source:
          $ref: "#/components/schemas/UserSource"
...
    AppointmentStatus:
      type: string
      enum: [scheduled, completed, cancelled]
    AppointmentSource:
      type: string
      enum: [llm, telegram_bot, web, admin]
    ServiceLineItem:
      type: object
      description: Позиция услуги в записи или визите
      required: [service_id, quantity]
      properties:
        service_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          default: 1
...
    Appointment:
      type: object
      required:
        - id
        - user_id
        - starts_at
        - ends_at
        - total_price
        - status
        - created_at
        - service_items
        - source
        - discount_percent
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        total_price:
          type: string
          description: "Расчётная стоимость на момент записи, десятичная строка"
          example: "4300.00"
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        created_at:
          type: string
          format: date-time
        service_items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        source:
          $ref: "#/components/schemas/AppointmentSource"
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
...
    AppointmentAdmin:
      allOf:
        - $ref: "#/components/schemas/Appointment"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
    AdminAppointmentListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentAdmin"
        total:
          type: integer
          description: Общее число записей без учёта пагинации
...
    Visit:
      type: object
      required:
        - id
        - appointment_id
        - total_amount
        - bonus_spent
        - bonus_earned
        - confirmed_at
        - confirmed_by_user_id
        - lines
      properties:
        id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        total_amount:
          type: string
          description: "Итоговая сумма оплаты, десятичная строка"
          example: "4300.00"
        bonus_spent:
          type: integer
        bonus_earned:
          type: integer
          description: Начислено бонусов по итогам визита (рассчитывает сервер)
        confirmed_at:
          type: string
          format: date-time
        confirmed_by_user_id:
          type: string
          format: uuid
        lines:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        client_rating_stars:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
        client_rating_comment:
          type: string
          nullable: true
        service_rating_stars:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
        service_rating_comment:
          type: string
          nullable: true
    VisitListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
          description: Общее число визитов без учёта пагинации
...
    DashboardTodayResponse:
      type: object
      required:
        - date
        - appointments_total
        - visits_total
        - cancellations_total
        - bookings_scheduled_today_by_source
        - consultation_user_messages_last_7_days
      description: |
        «Сегодня» — календарная дата в `consultation_business_timezone`.
        Записи на день считаются по **пересечению** интервала [starts_at, ends_at) с локальным
        полуинтервалом [00:00; 24:00) этого дня (как отображение в сетке недели), а не только
        по вхождению `starts_at` в день.
      properties:
        date:
          type: string
          format: date
          description: Локальная дата «сегодня» в TZ бизнеса
        appointments_total:
          type: integer
          description: Все записи, пересекающие этот день (любой статус)
        visits_total:
          type: integer
          description: Визиты с confirmed_at, попадающим в этот локальный день
        cancellations_total:
          type: integer
          description: Записи в статусе cancelled, пересекающие этот день
        bookings_scheduled_today_by_source:
          type: object
          description: |
            Число записей по полю source среди scheduled, completed и cancelled за день
            (то же правило пересечения с днём). Ключи ответа — `admin`, `web`, `llm`, `telegram_bot`;
            неизвестный source на сервере может агрегироваться в `web`.
          additionalProperties:
            type: integer
        consultation_user_messages_last_7_days:
          type: integer
          description: Сообщения пользователя в консультациях за последние 7 суток (UTC)
    WeekGridEvent:
      type: object
      required:
        - state
        - appointment_id
        - total_price
        - client_name
        - service_summaries
      properties:
        state:
          type: string
          enum: [scheduled, completed, cancelled]
        appointment_id:
          type: string
          format: uuid
        visit_id:
          type: string
          format: uuid
          nullable: true
        total_price:
          type: string
        client_name:
          type: string
        service_summaries:
          type: array
          items:
            type: string
        client_rating_stars:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
          description: Оценка клиента после визита; только для state=completed при наличии визита
        client_rating_comment:
          type: string
          nullable: true
    WeekGridSlot:
      type: object
      required: [starts_at, ends_at, events]
      properties:
        starts_at:
          type: string
          format: date-time
          description: Начало слота (UTC, ISO с Z)
        ends_at:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/WeekGridEvent"
    WeekGridDay:
      type: object
      required: [date, slots]
      properties:
        date:
          type: string
          format: date
        slots:
          type: array
          items:
            $ref: "#/components/schemas/WeekGridSlot"
    WeekGridResponse:
      type: object
      required: [week_start, slot_step_minutes, days]
      properties:
        week_start:
          type: string
          format: date
        slot_step_minutes:
          type: integer
        days:
          type: array
          items:
            $ref: "#/components/schemas/WeekGridDay"
    AnalyticsWeekResponse:
      type: object
      required: [week_start, top_services, days]
      properties:
        week_start:
          type: string
          format: date
        top_services:
          type: array
          items:
            type: object
            additionalProperties: true
        days:
          type: array
          items:
            type: object
            additionalProperties: true
    AdminDataInsightRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
    AdminDataInsightResponse:
      type: object
      required: [summary, sql_executed, columns, rows, truncated]
      properties:
        summary:
          type: string
        sql_executed:
          type: string
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        truncated:
          type: boolean
    AdminClientRow:
      type: object
      required:
        - user_id
        - name
        - visits_count
        - total_spent
        - bonus_balance
      properties:
        user_id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
          nullable: true
        telegram_id:
          type: integer
          format: int64
          nullable: true
        telegram_username:
          type: string
          nullable: true
        visits_count:
          type: integer
        total_spent:
          type: string
        bonus_balance:
          type: integer
        rating_avg:
          type: string
          nullable: true
    AdminClientListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminClientRow"
        total:
          type: integer
    AdminAppointmentCreateBody:
      type: object
      required: [user_id, starts_at, service_items]
      properties:
        user_id:
          type: string
          format: uuid
          description: Клиент (роль client)
        starts_at:
          type: string
          format: date-time
          description: Начало записи; не в прошлом (UTC), иначе STARTS_AT_IN_PAST
        service_items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
    AdminAppointmentPatchBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        service_items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
        source:
          $ref: "#/components/schemas/AppointmentSource"
    AdminVisitPatchBody:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        total_amount:
          type: string
          nullable: true
        bonus_spent:
          type: integer
          minimum: 0
          nullable: true
        bonus_earned:
          type: integer
          minimum: 0
          nullable: true
    VisitRatingBody:
      type: object
      required: [stars]
      properties:
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
...